data_producer/requirements.txt
pika==1.3.1
python-dotenv==0.21.0
pandas==1.5.2
pyarrow==10.0.1
pygelf==0.4.2